- Fixed a bug due to ljl::u8 being typedef'd to uint_least64_t. - Optimized __cdelc's prolog

+++ LJL/Aliases.h
@@ -7,9 +7,9 @@ LJL_BEGIN
 	typedef uint_least64_t u64;
 	typedef uint_least32_t u32;
 	typedef uint_least16_t u16;
-	typedef uint_least64_t u8 ;
+	typedef uint_least8_t  u8 ;
 	typedef  int_least64_t i64;
 	typedef  int_least32_t i32;
 	typedef  int_least16_t i16;
-	typedef  int_least64_t i8 ;
+	typedef  int_least8_t  i8 ;
 LJL_END
